Replace some calls to site.baseurl with absolute_url filter
Example: A stylesheet is located at /assets/styles.css ... On the
homepage (which is located at the root path of '/'), the href for that
stylesheet could be href="assets/styles.css", because that relative path
is LUCKILY valid:

homepage's path + relative href = /assets/styles.css

But on the Aboute page, the relative path is UNLUCKILY not valid. The
About page is located at /about/ so:

about's path + relative href = /about/assets/styles.css

That's not where the stylesheet actually lives.

So we need the paths to that stylesheet to NOT BE RELATIVE.

In Liquid, {{ "a string" | filter_method }} means "process the given
string with the filter method"... e.g. {{ "beth haubert" | capitalize }}
would return "Beth Haubert".

So I replaced site.baseurl usages with {{ "some/path/styles.css" |
absolute_url }}, which returns
"https://bethhaubert.com/some/path/styles.css" ... So it's no longer a
relative path at all. Whether you're looking at the homepage or the
About page, you will always be asking the same absolute path for
stylesheets.

---

Note: The reason for this change is that, for some unknown reason,
site.baseurl was returning nil (empty string, actually) on production.
It should be returning "/" which would have worked fine. But instead of
debugging that, I worked around the issue instead.